Added Xamarin Studio Plugin Project Added and empty project for a Xamarin Studio Plugin. After reading some of the documentation here: https://developer.xamarin.com/guides/cross-platform/xamarin-studio/customizing-ide/extending_xamarin_studio_with_addins/
it looks like it'll be easy to implement like for Visual Studio. All
that really needs to happen is we need an entry point, menu entry and
reference to the Class Library with the actual heavy-lifting code. --
Assuming all runs well. Currently, there is an intermixing of Windows
and Unix file path formats which is causing a crash. After the crash, we
may have difficulties with the Solution Parsing.